Patch Set 2/2: Add an "Automatically Connect" option for WiFi networks
Relies on: http://review.cyanogenmod.org/28529

Allows the ability for a user to determine whether to connect to an SSID automatically.
e.g. At work, there are several SSIDs with long, complex keys.
    I don't want to connect to them all of the time, but would like them to be remebered.
The default behaviour is to allow automatic connection, as it is today.

Patch set 2 : Fix capitalization & whitespace

Change-Id: I0f0aa64498115e3e1058f73113d37df63ab72d60