Implement saber mods:
Much of the ported features that have been ported from CyanogenMod
(about half), were implement/ported by me and not rascarlo.
Because of this I don't feel it represents what rascarlo himself has
in mind for features of his ROM builds.  But just for the record big
thanks to rascarlo for the following features ported from CyanogenMod
that will now be in saber mods:

  Clock - Show Clock, AM/PM style
  Battery status style
  Lockscreen shortcuts

Signed-off-by:  Paul Beeler <sparksco@gmail.com>