Add a option to hide power off option for power menu on lockscreen (2/2):
Some security apps require the device to be turned on to allow the
owner to have full remote control over the device in the event that
the device is stolen.  Having the power off option on lockscreen
can allow anyone to power off the device even when a lockscreen pin,
password or facelock is being used.

Signed-off-by: Paul Beeler <sparksco@gmail.com>